ref: remove duplicate function, set amp_factor as Uint64 variable when it's read-only
