+++ contracts/hub/vlp/src/contract.rs
@@ -164,7 +164,6 @@ pub fn execute(
             info,
             &STATE,
             &BALANCES,
-            None,
             sender,
             asset_in,
             amount_in,
