+++ contracts/hub/vlp/src/query.rs
@@ -1,10 +1,10 @@
 use cosmwasm_std::{
-    ensure, to_json_binary, Binary, Decimal, Decimal256, Deps, Env, Isqrt, Uint128, Uint256,
+    ensure, to_json_binary, Binary, Decimal, Decimal256, Deps, Env, Isqrt, Uint128,
 };
 use euclid::chain::ChainUid;
 use euclid::error::ContractError;
 use euclid::fee::BPS_50_PERCENT;
-use euclid::pool::{PoolConfig, MINIMUM_LIQUIDITY};
+use euclid::pool::{calculate_swap, PoolConfig, MINIMUM_LIQUIDITY};
 use euclid::swap::NextSwapVlp;
 use euclid::token::{Pair, PairWithAmount, Token};
 
@@ -184,27 +184,6 @@ fn get_pool(
             .unwrap_or(Uint128::zero()),
         lp_shares: chain_lp_tokens,
     })
-}
-// Function to calculate the asset to be recieved after a swap
-pub fn calculate_swap(
-    swap_amount: Uint128,
-    reserve_in: Uint128,
-    reserve_out: Uint128,
-) -> Result<Uint128, ContractError> {
-    let reserve_in = Uint256::from(reserve_in);
-    let reserve_out = Uint256::from(reserve_out);
-    // Calculate the k constant product
-    let k = reserve_in.checked_mul(reserve_out)?;
-    // Calculate the new reserve of token 1
-    let new_reserve_in = reserve_in.checked_add(swap_amount.into())?;
-    // Calculate the new reserve of token 2
-    let new_reserve_out = k.checked_div(new_reserve_in)?;
-    // Calculate the amount of token 2 to be recieved
-    let token_2_recieved = reserve_out.checked_sub(new_reserve_out)?;
-    let token_2_recieved =
-        Uint128::try_from(token_2_recieved).map_err(|_| ContractError::new("Overflow"))?;
-
-    Ok(token_2_recieved)
 }
 
 pub fn calculate_lp_allocation(
